ui, group, todo

+++ app/src/main/java/com/example/notecat/activities/EditNoteActivity.kt
@@ -5,6 +5,7 @@ import android.os.Bundle
 import android.text.SpannableStringBuilder
 import android.util.Log
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.ViewModelProvider
 import com.example.notecat.R
 import com.example.notecat.databinding.ActivityAddNoteBinding
@@ -19,7 +20,7 @@ class EditNoteActivity : AppCompatActivity() {
         binding = ActivityEditNoteBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        val imgv_cancel = binding.imgCancel
+        val imgv_delete = binding.imgDeletess
         val imgv_edit = binding.imgOkEdit
         val edt_title_edt = binding.edtTitleEdit
         val edt_content_edt = binding.edtContentEdit
@@ -30,25 +31,31 @@ class EditNoteActivity : AppCompatActivity() {
             edt_content_edt.setText(note.content)
         }
         val noteViewModel = ViewModelProvider(this)[NoteViewModel::class.java]
-        imgv_cancel.setOnClickListener {
+
+        imgv_delete.setOnClickListener {
             if (note != null) {
-                noteViewModel.deleteNoteVM(note)
-                Toast.makeText(this,"Xóa thành công !", Toast.LENGTH_SHORT).show()
+                AlertDialog.Builder(this)
+                    .setTitle("XÓA GHI CHÚ")
+                    .setMessage("Bạn có muốn xóa?")
+                    .setPositiveButton("Xóa") { _, _ ->
+                        noteViewModel.deleteNoteVM(note)
+                        Toast.makeText(this, "Xóa thành công !", Toast.LENGTH_SHORT).show()
+                        finish()
+                    }
+                    .setNegativeButton("Hủy") { _, _ ->
+                        finish()
+                    }
+                    .show()
             }
-            finish()
         }
         imgv_edit.setOnClickListener {
             val newTitle = edt_title_edt.text
             val newContent = edt_content_edt.text
             if (note != null) {
                 note.title = newTitle.toString()
-            }
-            if (note != null) {
                 note.content = newContent.toString()
-            }
-            if (note != null) {
                 noteViewModel.updateNoteVM(note)
-                Toast.makeText(this,"Đã update !", Toast.LENGTH_SHORT).show()
+                Toast.makeText(this, "Đã update !", Toast.LENGTH_SHORT).show()
             }
 
             finish()
